+++ app/models/article.js
@@ -45,7 +45,7 @@ export default class ArticleModel extends Model {
   }
 
   async favoriteOperation(operation) {
-    let { article } = await this.authorizedFetch.fetch(
+    let { article } = await this.session.fetch(
       `/articles/${this.id}/favorite`,
       operation === 'unfavorite' ? 'DELETE' : 'POST',
     );
